Cleaning up seralization files
