+++ packages/wallet/src/serialization/abi-parsers.ts
@@ -1,12 +1,6 @@
-/* External Imports */
-import { ethers } from 'ethers'
-
-/* Internal Imports */
 import {
   CreateAndTransferTransition,
   InvalidTokenTypeError,
-  isSwapTransaction,
-  isTransferTransaction,
   SignedTransaction,
   Swap,
   SwapTransition,
@@ -16,8 +10,15 @@ import {
   TransferTransition,
 } from '../types'
 import { PIGI_TOKEN_TYPE, UNI_TOKEN_TYPE } from '../index'
-
-export const abi = new ethers.utils.AbiCoder()
+import {
+  abi,
+  createAndTransferTransitionAbiTypes,
+  signedTransactionAbiTypes,
+  swapAbiTypes,
+  swapTransitionAbiTypes,
+  transferAbiTypes,
+  transferTransitionAbiTypes,
+} from './common'
 
 const getTokenType = (tokenType): TokenType => {
   const tokenTypeNumber: number = +tokenType
@@ -29,8 +30,6 @@ const getTokenType = (tokenType): TokenType => {
   }
   return tokenTypeNumber
 }
-
-const transferAbiTypes = ['address', 'address', 'bool', 'uint32']
 
 /**
  * Creates a Transfer from an ABI-encoded Transfer.
@@ -50,19 +49,6 @@ export const parseTransferFromABI = (abiEncoded: string): Transfer => {
   }
 }
 
-/**
- * @returns the ABI-encoded Transfer.
- */
-export const abiEncodeTransfer = (transfer: Transfer): string => {
-  return abi.encode(transferAbiTypes, [
-    transfer.sender,
-    transfer.recipient,
-    transfer.tokenType,
-    transfer.amount,
-  ])
-}
-
-const swapAbiTypes = ['address', 'bool', 'uint32', 'uint32', 'uint']
 /**
  * Creates a Swap from an ABI-encoded Swap.
  * @param abiEncoded The ABI-encoded Swap.
@@ -84,20 +70,6 @@ export const parseSwapFromABI = (abiEncoded: string): Swap => {
 }
 
 /**
- * @returns the abi encoded Swap.
- */
-export const abiEncodeSwap = (swap: Swap): string => {
-  return abi.encode(swapAbiTypes, [
-    swap.sender,
-    swap.tokenType,
-    swap.inputAmount,
-    swap.minOutputAmount,
-    swap.timeout,
-  ])
-}
-
-const signedTransactionAbiTypes = ['bytes', 'bytes']
-/**
  * Creates a SignedTransaction from an ABI-encoded SignedTransaction.
  * @param abiEncoded The ABI-encoded SignedTransaction.
  * @returns the SignedTransaction.
@@ -112,29 +84,6 @@ export const parseSignedTransactionFromABI = (
     transaction: parseTransactionFromABI(tx),
   }
 }
-
-/**
- * @returns the abi encoded SignedTransaction.
- */
-export const abiEncodeSignedTransaction = (
-  signedTransaction: SignedTransaction
-): string => {
-  return abi.encode(signedTransactionAbiTypes, [
-    signedTransaction.signature,
-    abiEncodeTransaction(signedTransaction.transaction),
-  ])
-}
-
-const swapTransitionAbiTypes = [
-  'bytes32',
-  'uint32',
-  'uint32',
-  'bool',
-  'uint32',
-  'uint32',
-  'uint',
-  'bytes',
-]
 
 /**
  * Creates a SwapTransition from an ABI-encoded SwapTransition.
@@ -166,31 +115,6 @@ export const parseSwapTransitionFromABI = (encoded: string): SwapTransition => {
 }
 
 /**
- * @returns the ABI-encoded SwapTransition.
- */
-export const abiEncodeSwapTransition = (trans: SwapTransition): string => {
-  return abi.encode(swapTransitionAbiTypes, [
-    trans.stateRoot,
-    trans.senderSlot,
-    trans.recipientSlot,
-    trans.tokenType,
-    trans.inputAmount,
-    trans.minOutputAmount,
-    trans.timeout,
-    trans.signature,
-  ])
-}
-
-const transferTransitionAbiTypes = [
-  'bytes32',
-  'uint32',
-  'uint32',
-  'bool',
-  'uint32',
-  'bytes',
-]
-
-/**
  * Creates a TransferTransition from an ABI-encoded TransferTransition.
  * @param encoded The ABI-encoded TransferTransition.
  * @returns the TransferTransition.
@@ -218,32 +142,6 @@ export const parseTransferTransitionFromABI = (
 }
 
 /**
- * @returns the ABI-encoded TransferTransition.
- */
-export const abiEncodeTransferTransition = (
-  trans: TransferTransition
-): string => {
-  return abi.encode(transferTransitionAbiTypes, [
-    trans.stateRoot,
-    trans.senderSlot,
-    trans.recipientSlot,
-    trans.tokenType,
-    trans.amount,
-    trans.signature,
-  ])
-}
-
-const createAndTransferTransitionAbiTypes = [
-  'bytes32',
-  'uint32',
-  'uint32',
-  'address',
-  'bool',
-  'uint32',
-  'bytes',
-]
-
-/**
  * Creates a CreateAndTransferTransition from an ABI-encoded CreateAndTransferTransition.
  * @param encoded The ABI-encoded CreateAndTransferTransition.
  * @returns the CreateAndTransferTransition.
@@ -269,37 +167,6 @@ export const parseCreateAndTransferTransitionFromABI = (
     amount,
     signature,
   }
-}
-
-/**
- * @returns the ABI-encoded CreateAndTransferTransition.
- */
-export const abiEncodeCreateAndTransferTransition = (
-  trans: CreateAndTransferTransition
-): string => {
-  return abi.encode(createAndTransferTransitionAbiTypes, [
-    trans.stateRoot,
-    trans.senderSlot,
-    trans.recipientSlot,
-    trans.createdAccountPubkey,
-    trans.tokenType,
-    trans.amount,
-    trans.signature,
-  ])
-}
-
-/**
- * ABI-encodes the provided Transaction.
- * @param transaction The transaction to AbI-encode.
- * @returns The ABI-encoded Transaction as a string.
- */
-export const abiEncodeTransaction = (transaction: Transaction): string => {
-  if (isSwapTransaction(transaction)) {
-    return abiEncodeSwap(transaction)
-  } else if (isTransferTransaction(transaction)) {
-    return abiEncodeTransfer(transaction)
-  }
-  throw Error(`Unknown transaction type: ${JSON.stringify(transaction)}`)
 }
 
 /**
