+++ packages/wallet/src/serialization/index.ts
@@ -1,3 +1,2 @@
-import { ethers } from 'ethers'
-
-export * from './serialization'
+export * from './abi-parsers'
+export * from './abi-encoders'
